requirements.txt
fastapi
pydantic
uvicorn
pillow
opencv-python
selenium
webdriver-manager
requests
urllib3
numpy
vidgear
packaging